ai_system/requirements.txt
# Python dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
slack-sdk==3.26.2
requests==2.31.0
openai==1.12.0

